code refactoring

+++ Graph.h
@@ -33,7 +33,7 @@ public:
 private:
     float getDistance(int from, int to);
     vector<vector<Edge>> primMST();
-    void tspBackTracking(vector<bool> &v, int currPos, int n, int count, float cost, float &ans, vector<int> &path, vector<int> &bestPath);
+    void tspBackTracking(vector<bool> &v, int currPos, size_t n, int count, float cost, float &ans, vector<int> &path, vector<int> &bestPath);
     int input_vertex(const string &input_name);
     int input_edge(const string &input_name, bool have_nodes);
 };
